+++ src-python/qiskit_qasm2/__init__.py
@@ -5,8 +5,9 @@ from typing import Iterable, Union, Optional, Literal
 with open(Path(__file__).parent / "VERSION", "r", encoding="utf-8") as _version_file:
     __version__ = _version_file.read().strip()
 
-from . import core, parse
+from . import core as _core, parse as _parse  # pylint: disable=no-name-in-module
 from .core import QASM2ParseError
+from .parse import CustomInstruction, QISKIT_CUSTOM_INSTRUCTIONS
 
 
 def _normalize_path(path: Union[str, os.PathLike]) -> str:
@@ -21,7 +22,11 @@ def _normalize_path(path: Union[str, os.PathLike]) -> str:
     return str(path)
 
 
-def loads(string: str, include_path: Iterable[Union[str, os.PathLike]] = (".",)):
+def loads(
+    string: str,
+    include_path: Iterable[Union[str, os.PathLike]] = (".",),
+    custom_instructions: Iterable[CustomInstruction] = (),
+):
     """Parse an OpenQASM 2 program from a string into a :class:`~qiskit.circuit.QuantumCircuit`.
 
     :param string: The OpenQASM 2 program in a string.
@@ -30,8 +35,17 @@ def loads(string: str, include_path: Iterable[Union[str, os.PathLike]] = (".",))
     :return: A circuit object representing the same OpenQASM 2 program.
     :rtype: ~qiskit.circuit.QuantumCircuit
     """
-    return parse.from_bytecode(
-        core.bytecode_from_string(string, [_normalize_path(path) for path in include_path])
+    custom_instructions = list(custom_instructions)
+    return _parse.from_bytecode(
+        _core.bytecode_from_string(
+            string,
+            [_normalize_path(path) for path in include_path],
+            [
+                _core.CustomInstruction(x.name, x.n_params, x.n_qubits, x.builtin)
+                for x in custom_instructions
+            ],
+        ),
+        custom_instructions,
     )
 
 
@@ -39,6 +53,7 @@ def load(
     filename: Union[str, os.PathLike],
     include_path: Iterable[Union[str, os.PathLike]] = (".",),
     include_input_directory: Optional[Literal["append", "prepend"]] = "append",
+    custom_instructions: Iterable[CustomInstruction] = (),
 ):
     """Parse an OpenQASM 2 program from a file into a :class:`~qiskit.circuit.QuantumCircuit`.  The
     given path should be ASCII or UTF-8 encoded, and contain the OpenQASM 2 program.
@@ -65,4 +80,15 @@ def load(
             f"unknown value for include_input_directory: '{include_input_directory}'."
             " Valid values are '\"append\"', '\"prepend\"' and 'None'."
         )
-    return parse.from_bytecode(core.bytecode_from_file(_normalize_path(filename), include_path))
+    custom_instructions = tuple(custom_instructions)
+    return _parse.from_bytecode(
+        _core.bytecode_from_file(
+            _normalize_path(filename),
+            include_path,
+            [
+                _core.CustomInstruction(x.name, x.n_params, x.n_qubits, x.builtin)
+                for x in custom_instructions
+            ],
+        ),
+        custom_instructions,
+    )